refactor: 폴더 구조 개선 (#34)
- src/comment/ → src/api/comment/, src/hooks/comment/ 분리
- API 레이어(services)와 비즈니스 로직(queries) 분리
- import 경로 업데이트
  - queries.ts: @/comment/services → @/api/comment/services
  - usecomments.ts: @/comment/queries → @/hooks/comment/queries
  - useCommentStore.ts: @/comment/services → @/api/comment/services
